Create traits for pagination and remove some code
