+++ app/Traits/Pagination.php
@@ -1,0 +1,30 @@
+<?php
+
+namespace App\Traits;
+
+use Illuminate\Support\Collection;
+use Illuminate\Support\Facades\Validator;
+use Illuminate\Pagination\LengthAwarePaginator;
+
+trait Pagination
+{
+    protected function paginate(Collection $collection)
+    {
+        $page = LengthAwarePaginator::resolveCurrentPage();
+
+        $perPage = 2;
+        if(request()->has('per_page')){
+            $perPage = (int)request()->per_page;
+        }
+
+        $results = $collection->slice(($page-1) * $perPage, $perPage)->values();
+
+        $paginated = new LengthAwarePaginator($results, $collection->count(), $perPage, $page, [
+            'path' => LengthAwarePaginator::resolveCurrentPage()
+        ]);
+
+        $paginated->appends(request()->all());
+
+        return $paginated;
+    }
+}
