+++ app/Traits/ResponseMessage.php
@@ -33,36 +33,4 @@ trait ResponseMessage
     {
         return $this->successResponse($array, $code);
     }
-    protected function paginatedList(Collection $collection, $code = 200)
-    {
-        $collection = $this->paginate($collection);
-
-        return $this->successResponse(['data' => $collection], $code);
-    }
-
-    protected function paginate(Collection $collection)
-    {
-        $rules = [
-            'per_page' => 'integer|min:2|max:50'
-        ];
-
-        Validator::validate(request()->all(), $rules);
-
-        $page = LengthAwarePaginator::resolveCurrentPage();
-
-        $perPage = 12;
-        if(request()->has('per_page')){
-            $perPage = (int)request()->per_page;
-        }
-
-        $results = $collection->slice(($page-1) * $perPage, $perPage)->values();
-
-        $paginated = new LengthAwarePaginator($results, $collection->count(), $perPage, $page, [
-            'path' => LengthAwarePaginator::resolveCurrentPage()
-        ]);
-
-        $paginated->appends(request()->all());
-
-        return $paginated;
-    }
 }
